Tighten remaining astra-plan surface
