+++ rust/crates/astra-plan/src/lib.rs
@@ -6,5 +6,6 @@ pub mod repository;
 pub use decompose::*;
 pub use repository::{
     CloudPlanRepository, InMemoryPlanRepository, NewStepRun, PlanListFilter, PlanLoadError,
-    PlanRepository, PlanStepRun, plan_resume_hint_for_session, plan_resume_prompt_hint,
+    PlanRepository, PlanStepRun, SavedPlanInfo, plan_resume_digest, plan_resume_hint_for_session,
+    plan_resume_prompt_hint,
 };
